chore: move configuration property audit listener
Refs: XRDDEV-3192
